services/auth-service/requirements.txt
asgiref==3.11.0
Django==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
pillow==12.0.0
PyJWT==2.10.1
python-dotenv==1.2.1
sqlparse==0.5.5
tzdata==2025.3
django-cors-headers==4.3.1
pika==1.3.2
psycopg[binary]


